# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Risk & Disputes API
  version: 0.2.0
  description: Risk engine, fraud signals, dispute management, evidence, and representment
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3003
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/risk/signals:
    get:
      summary: List risk signals
      description: Returns a paginated list of risk signals for the merchant.
      x-platform:
        kind: resource
        entity: RiskSignal
        affectsLedger: false
      operationId: listRiskSignals
      tags:
        - RiskSignals
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of risk signals.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a risk signal
      description: Ingests a new risk signal for a given entity.
      x-platform:
        kind: resource
        entity: RiskSignal
        emitsEvents:
          - risk_signal.created
        affectsLedger: false
      operationId: createRiskSignal
      tags:
        - RiskSignals
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entity_type
                - entity_id
                - signal_type
                - severity
              properties:
                entity_type:
                  type: string
                entity_id:
                  type: string
                signal_type:
                  type: string
                severity:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                    - critical
      responses:
        "201":
          description: Risk signal created.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...